+++ pybgfx/__init__.py
@@ -1,0 +1,9 @@
+from .initializor import initialise
+from .constants import *
+
+# FIXME: different platforms
+initialise('bgfx', 'libbgfx_pythonCppyy.dylib', 'bgfx_python.map')
+del initialise
+
+
+
