+++ pybgfx/constants.py
