--- scripts/all_cmake_includes_bgfx.hpp
@@ -1,3 +1,0 @@
-#include <bgfx.h>
-#include <platform.h>
-#include <shaderc.h>
