--- scripts/all_cmake_includes_imgui.hpp
@@ -1,2 +1,0 @@
-#include <imgui.h>
-#include <imgui_internal.h>
