--- scripts/config.cfg
@@ -1,15 +1,0 @@
-+include <pybind11/stl.h>
-
-+namespace bgfx
-+class bgfx::Stats
-+class bgfx::EncoderStats
-+class bgfx::ViewStats
-+class ImGuiListClipper
-
-+default_static_pointer_return_value_policy           pybind11::return_value_policy::reference
-+default_static_lvalue_reference_return_value_policy  pybind11::return_value_policy::reference_internal
-+default_static_rvalue_reference_return_value_policy  pybind11::return_value_policy::move
-+default_member_pointer_return_value_policy           pybind11::return_value_policy::reference
-+default_member_lvalue_reference_return_value_policy  pybind11::return_value_policy::reference_internal
-+default_member_rvalue_reference_return_value_policy  pybind11::return_value_policy::move
-+default_call_guard                                   pybind11::gil_scoped_release
