--- src/shaderc_module.cpp
@@ -1,34 +1,0 @@
-
-#include <map>
-#include <memory>
-#include <stdexcept>
-#include <functional>
-#include <string>
-
-#include <pybind11/pybind11.h>
-
-typedef std::function< pybind11::module & (std::string const &) > ModuleGetter;
-
-void bind_shaderc(std::function< pybind11::module &(std::string const &namespace_) > &M);
-
-PYBIND11_MODULE(shaderc, root_module) {
-	root_module.doc() = "shaderc module";
-
-	std::map <std::string, pybind11::module> modules;
-	ModuleGetter M = [&](std::string const &namespace_) -> pybind11::module & {
-		auto it = modules.find(namespace_);
-		if( it == modules.end() ) throw std::runtime_error("Attempt to access pybind11::module for namespace " + namespace_ + " before it was created!!!");
-		return it->second;
-	};
-
-	modules[""] = root_module;
-
-	std::vector< std::pair<std::string, std::string> > sub_modules {
-	    {"", "shaderc"},
-	};
-	for(auto &p : sub_modules ) modules[p.first.size() ? p.first+"::"+p.second : p.second] = modules[p.first].def_submodule(p.second.c_str(), ("Bindings for " + p.first + "::" + p.second + " namespace").c_str() );
-
-	//pybind11::class_<std::shared_ptr<void>>(M(""), "_encapsulated_data_");
-
-    bind_shaderc(M);
-}
